Rewrite listener for parsing of incoming messages
Listener now properly adds incoming messages to the Groups in a game, so
that a game can then decid whether to delete these or not.
Minor improvements in naming of Groups, minor fixes.
Now ready to implement settings for which messages should be deleted

Squashed commit of the following:

commit 070b0bdc566a078ceb9b1a7479d5b3c26c9d8d1f
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:47:34 2021 +0200

    fix error

commit 7b891c070e613cef7b7e95a3c6a0cc949c16f75a
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:47:15 2021 +0200

    correct to new name

commit 8b1fd5f3afc793c05e71df5c652d3e66ef938c2d
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:46:54 2021 +0200

    Add default clearing exceptions

commit a68550c3795491405e7f780c933bfa022b393af2
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:43:14 2021 +0200

    Rename groups for more direct group names

commit c7206925ee673dc81ce22fd396c1e570ab4020a3
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:42:43 2021 +0200

    Rewrite the listener to properly process incoming messages and put them into the corresponding Groups / add hints

commit 0ad33c843ac717ab35bd8227e51ee114ad7ec61a
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:31:34 2021 +0200

    fix: remove wrong line (still there from copy, but not needed)

commit b8b6390c368e6c8e87c313d76db0609232fe3344
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:21:39 2021 +0200

    make rules command use the proper message from output.py

commit 3d04834d1272f865037d8334b9b138b5629a8441
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:19:05 2021 +0200

    move rules method to output.py

commit 49b6895531a05d81bfa09e1da7b8565a79058d42
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:18:40 2021 +0200

    add todo

commit f106958815665e1dbcac98816de0d19bf653afe8
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:18:27 2021 +0200

    Preserve certain word lists when clearing (default)

commit 085c2da71d64f96e3b4bb93e7196d237954965ce
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:10:20 2021 +0200

    Fix double enumeration number of summary

commit 6f5367a49ba1317b57f75ea432e1cd8bd1214ce7
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:09:54 2021 +0200

    Add exception handling in case channel has been deleted when sending a warn message

commit b386a738f207bb8a6233b6566a5d7e7f12e34c24
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:09:17 2021 +0200

    provide options to preserve certains groups or messages when clearing

commit 2f67b069356474e6ae3159d2b55a4370e4c9d6a7
Author: Maximilian Keßler <max@maximilian-kessler.de>
Date:   Tue May 25 17:08:36 2021 +0200

    Fix summary message not being indexed by key

+++ src/game_management/messages.py
@@ -80,12 +80,18 @@ class MessageHandler:  # Basic message handler for messages that one wants to se
         if pop:
             self.special_messages.pop(key)
 
-    async def clear_all(self):
+    async def clear_all(self, preserve_keys: List[Key] = [], preserve_groups: List[Group] = []):
+        print(self.group_messages)
         for group_key in self.group_messages.keys():
-            await self.delete_group(group_key)
-        for special_message_key in self.special_messages.keys():
-            await self.delete_special_message(special_message_key, pop=False)
-        self.special_messages = {}
+            if group_key not in preserve_groups:
+                await self.delete_group(group_key)
+
+        print(f"Special messages (dict): {self.special_messages}\n")
+        special_message_keys = [special_message_key for special_message_key in self.special_messages.keys()]
+        for special_message_key in special_message_keys:
+            print(f"Checking key {special_message_key} within list {preserve_keys}")
+            if special_message_key not in preserve_keys:
+                await self.delete_special_message(special_message_key, pop=True)
 
     async def wait_for_reaction_to_message(self,
                                            bot: discord.ext.commands.Bot,
@@ -131,8 +137,12 @@ class MessageHandler:  # Basic message handler for messages that one wants to se
             return True  # Notify that reaction was found
         except asyncio.TimeoutError:
             print('No reaction, send warning')
-            await self.message_sender.send_message(normal_text=f"Hey, {member.mention if member else ''}",
+            # TODO check if warning is appropiate, i.e. if the original message still exists. Else, abort this
+            try:
+                await self.message_sender.send_message(normal_text=f"Hey, {member.mention if member else ''}",
                                                    embed=warning, reaction=False, channel=message.channel, group=Group.warn)
+            except discord.NotFound:  # In case the channel does not exist anymore
+                return False
             # Try a second time
             try:
                 reaction, user = await bot.wait_for('reaction_add', timeout=timeout, check=check)
